+++ src/Bible.Alarm/Common/Interfaces/Media/IMediaCacheSetupService.cs
@@ -2,6 +2,6 @@ namespace Bible.Alarm.Common.Interfaces.Media;
 
 public interface IMediaCacheSetupService
 {
-    Task SetupAlarmCacheAsync(long scheduleId);
+    Task SetupAlarmCacheAsync(int scheduleId);
 }
