+++ src/Bible.Alarm/Services/Scheduler/IScheduleStateService.cs
@@ -2,6 +2,6 @@ namespace Bible.Alarm.Services.Scheduler;
 
 public interface IScheduleStateService
 {
-    Task<bool> UpdateScheduleEnabledStateAsync(long scheduleId, bool isEnabled);
+    Task<bool> UpdateScheduleEnabledStateAsync(int scheduleId, bool isEnabled);
 }
